readyset-dataflow: Bottom level barrier forwarding during eviction
If an eviction message contains a barrier, we create an object to
track the barrier's state.  During eviction, we stick outgoing
messages into a special holding pen.  At the end of the eviction,
we split received barrier credits among the pending, outgoing
eviction messages, or, if there are no further downstream evictions,
we return the credits to the worker.

Aside from the unit tests, the code is not yet active.

Change-Id: Ifd9ab83e092523d2b5fc4c960589005dac484f6c