readyset-server: Delete redundant, nonsensical impl Sync
(0) While skimming, I noticed an impl Sync, which made my ears perk
up a bit.  Fortunately, it was commented.

(1) The comment talked a lot about Send, then implemented Sync.  Huh?

(2) I tried to follow the chain of types it mentioned, and the comment
appears to have gone stale.

(3) I deleted the impl, and everything typechecked fine.

(4) I checked whether the type is auto Send and Sync, and it is.
Apparently the error on duplicate impls isn't triggered if you unsafe
impl an auto trait that was already autoimplemented.

Let's delete this lurking land mine.

Change-Id: Ifedbfb049e0ba4752f2d6fea1e79dcceb41178e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7880
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>