readyset-dataflow: Log failed base table writes
We were previously intending to return an error when detecting writes,
but an inadvertently flipped boolean meant we would only detect errors
when in "permissive" mode (no upstream). We would like to resurrect that
error detection, but as it is this causes the domain to fail, which
causes more problems than it solves.

Co-authored-by: Johnathan Davis <jcd@readyset.io>
Change-Id: Ie189cd7a5c89dfd02bef1f3c667dc8c2fd1836e9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7894
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>