dataflow-state: Use flush_wal function to sync the WAL
Part of our fork of RocksDB exposed a separate function to sync the WAL
only and then did that after a flush_wal(false) that wrote from memory
to disk (without syncing).  But flush_wal(true) is practically speaking
going to be just as good and doesn't require forking a dependency.

Change-Id: Iea1231e9c500ad8c41cfd9dc63b56540affcbb89