readyset-server: Decrease aggressive eviction aggressiveness
During an unrelated buildkite run, I got an RPC error that said "request
timed out." I looked for the timeout, but I couldn't figure out what it
is or where it's set.  Since I haven't heard of any other failures, the
test is probably only very slightly marginal, so let's decrease the
frequency slightly from "totally ludicrous" to a mere "fairly nuts" in
the hope that the system will be able to survive at that rate.

Change-Id: I2e4aa4308f06d4829764a5e208f115683185435d