replicators: fix handling of JSON serde to text conversion
This commit fixes the handling of JSON serde to text conversion in the
MySQL replicator. The previous implementation was handling json
conversion to serde_json only on replication, but not during snapshot.
On snapshot we were relying on MySQL json format. This causes some
updates/deletes to not find the correct row because the order in which
the fields are serialized might be different.

Now we are using the same serde_json format for both replication and
snapshot to ensure the correctness of the data.

Change-Id: Ic3a04c310ed7470a95a637c0397b4c92321e7cf5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7910
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>