reader-map: Delete impls of Deref and AsRef
Both of these implementations are leaking the internal storage mechanism
of the Values type, which prevents us from changing it.  In particular,
the implementation Deref borders on abusive since that trait is intended
for smart pointers, which Values is not, and is just proxying function
calls to the type's inner storage.

Change-Id: Iaa2608553a53412c4ba9b31eefe1b964444db891
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7954
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI