dataflow-state: Handle non-existent deployment dirs
At process start up, Readyset will clean out the tmp/working dir,
in case the process crashed on a previous run and we need to delete
temp files. If the `working_dir` already exists by the time the
`clean_working_dir()` function is invoked, then everything is fine. If
the directory does not exist, then the call to `Path::canonicalize()`
will fail.

This problem does not exist in standalone, as the Standalone Authority
is created (and it's storage directories) much earlier than we attempt
to clean out the working dir. This is not the case in adapter/server
operation, which is where this was failing. The fix is to simply not
try to delete or canonicalize the working directory path if the
directory does not exist.

Release-Note-Core: Fix a bug in distributed readyset where
  readyset-server will fail on startup due to a missing working_dir.
Change-Id: Ibe7b0511d285b224a36ce7568673105926f1c354
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7920
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>