clustertest: Build dependencies before running
This test used to typically fail by default when run on a dev machine.

It depends on readyset and readyset-server binaries built in release
mode and with the failure_injection feature.  To get the test to run
correctly, you would first need to compile those binaries manually and
tell readyset-clustertest where they are with the BINARY_PATH
environment variable.

Instead of all that, we now use a build.rs file to automatically build
the test's dependencies when the test is run via cargo.  The binaries
are placed in a clustertest target directory to avoid forcing recompiles
in the regular target directory and also to prevent a deadlock that
cargo would otherwise experience if our recursive cargo invocation were
to try to grab the same build directory lock that the first cargo
process is already holding.

Now, devs should be able to run this test simply with:

  cargo test -p readyset-clustertest

Change-Id: I5ec3297a8fa4ed4abb86c8e074ddcb6f7ccee1c0