Switch toolchain from nightly to beta-2024-09-05
After 51 corrective patches, we are preparing to switch to the stable
Rust toolchain.  At this time, rotate your launch key to SET.

The use of unstable compiler features is now prohibited.

- Switch to new style toolchain config file
- Adjust infra (mostly Dockerfiles) to the new filename
- Install rust-analyzer automatically on toolchain bump
- Replace duplicate toolchain/rustfmt configs with symlinks
- Delete remaining feature gates
- Minor clippy fixes
- Delete comment about manually installing toolchain, which cargo does

We disable a new clippy lint on the length of the first paragraph of doc
comments.  The error is totally unhelpful (it says it's "too long"), and
it doesn't say the maximum length it will deign to permit.  I had to
read the commit on Github to learn that the limit is hardcoded to 200
characters, but they don't count rendering, so Markdown links count
against you.  I started fixing them, and while I get the point, I think
the value for us is low since our doc comments aren't public
documentation that people are regularly reading.

Change-Id: I5ab3e5b64c6805c3731cc0eb60ed40406d979997
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7927
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Ron Hough <ron@readyset.io>