readyset-server: Further decrease aggressive eviction aggressiveness
Looking at a recent failure, I see that the timeout isn't from eviction
but rather from a view request.  The default timeout appears to be five
seconds, which sounds like a lot, but this is running during a bunch of
simultaneous tests, so who knows what kind of performance we can expect.
We crank that timeout and further decrease the degree of eviction
insanity, which nevertheless remains somewhat insane.

Change-Id: I9e52e219832261cf10d1fa90d4c9aeceeb60088e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7962
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>