clustertest: Don't build dependencies if only running clippy
When cargo clippy is run, this build.rs program is also executed, which
can potentially add minutes to a simple invocation of cargo clippy.

Use the presence of the CLIPPY_ARGS environment variable to determine
whether or not we should continue building the clustertest dependencies.

If CLIPPY_ARGS is set, we presume that this invocation of cargo was from
cargo clippy and skip the compilation of the clustertest dependencies.

Change-Id: I3a9b4754ba637af81885473f7af58c773df11e9e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7981
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>