dataflow-expression: Add support for the length's functions
This commit adds support for the length, char_length, character_length
and octet_length functions. These functions behavior depend on their
dialect. PostgreSQL counts amount of bytes when using octet_length and
characters otherwise. For MySQL, length and octet_length is the number
 of bytes in the string, and char_length or character_length is the
number of characters in the string.
Also, PostgreSQL is strict to on the input type, so it will return an
error if the input is not a string. MySQL, on the other hand, will
convert the input to a string and return the length of the string.

Release-Note-Core: Add support for `length`, `char_length`,
  `character_length` and `octet_length` functions.
Change-Id: I888f1e8bfde1c61cd4a00a66fd40727a8f344bc1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8004
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>