reader-map: Delete generational eviction
We currently have three mediocre ways of evicting here, only one of
which is used.  Now, we shall have two mediocre ways.  Some day, we
shall have one way that is good.

The generational code is similar to the LRU code, but the generations
are the time between received evictions, which could be quite long, and
within the last generation, we evict randomly.  Neither of these are
things we want in eviction.

Change-Id: I89f61b175b4779c90e2c2d5d304a6f0a8bcba79f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8049
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>