Disable Post Lookup by default
It turns out there is a usage with LIMIT and JOIN that we need to take a
look at before enabling this broadly.

This is also a partial revert of 44add155.  We again pass the argument
explicitly as we did before the change that made this feature enabled by
default.

Addresses: REA-4797
Change-Id: I52ecc645df6bd34b5b03f43f376ab4939730213e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8011
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>
Tested-by: Buildkite CI