Disable Straddled Joins by default
Out of an abundance of caution, we are also disabling this by default
for now.  We don't, however, have any examples of issues with this
feature.

This is also a partial revert of Iad455b26.  We again pass the argument
explicitly as we did before the change that made this feature enabled by
default.

Change-Id: I5b79cf656e5e161ba9e2d7c3013541d6d560a7d9