reader-map: Change eviction meta u64 to from counter to clock
Since we will be fine with approximate LRU, we don't need to use a
counter that is precisely maintained via expensive atomics.  Instead,
switch to a coarse clock.

(Though the store at relaxed consistency may contend on the cache line,
disassembly of the binary shows that it's not actually atomic, i.e. not
actually locking the bus, despite what it may say on the tin.)

Change-Id: I64f39bef081833d9f67c0b278460f52f583d7a29
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8052
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>