reader-map: Switch from std HashMap to indexmap crate
The crate provides a superset of the standard container's functionality.
The additional part we want is its ability to select a hash table entry
by index, which will allow us to sample entries efficiently.  (The only
way we can sample the standard HashMap is by iterating it, which
requires us to iterate over all entries, which we don't want to do.)

The crate claims similar performance, though I observed a c. 5% speedup
in the product when running a simple end-to-end workload.

Change-Id: I43bc4309aa4c4f6f81f4e0dcd9ae31609840bec1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8054
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>