reader-map: Refactor eviction tests
We pull them into the crate as unit tests so that they can access
crate-private functionality, which integration tests cannot.  We also
significantly rewrite the tests to make them check for a general
improvement instead of checking for the eviction of specific keys.
Some redundant testing relating to intervals was removed, which is
covered by other tests, and which required specific, predictable key
eviction.

Change-Id: I5a6166958bd316d8240696f5da627fd13a09e58d