reader-map: Make LRU eviction faster and more approximate
Testing on a certain system revealed that with 170k keys in a reader, we
were spending 2.1 ms deciding which keys to evict.  That seemed like too
much.  Rather than partitioning all timestamps to determine a precise
cutoff value, we now sample a sublinear fraction of them and partition
the sample.  In the same test, cutoff selection now takes about 5 μs,
about a 400x improvement.

The tradeoff is increased error in the actual size of the eviction
relative to the desired size, which in testing was observed to be
occasionally as high as 50%.  But if we underevict, we'll just evict
again soon, so it's no big deal.

Change-Id: I7330a771400262c6ece393419d8caf54a893c6d8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8056
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>