replicators: pass table filters as reference
Our replicator code (noria_adapter) has a bug with regards to the
table filter. When we start the database replicator(which despite the
name, only does snapshot and not replication), we pass a copy of
the table_filter. During snapshot, we might fail to parse a table, and
in that case, we will add this table to the non replication tables.
The noria adapter is responsible for handling the replication events
after the snapshot has being completed. If we add a table to the
non-replication tables during snapshot, since we passed a copy of the
table_filter, the table_filter in the noria_adapter will not be
updated.
Adjusting it to pass a reference of table filter to replicators. This
is safe as while snapshot is running, the noria adapter has not
started replicating events yet.

Change-Id: Ia49ec282a20bc10a974253f598170b3f3ce9c7d7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8106
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>