noria-adapter: add support for snapshotting new tables
This change adds support for snapshotting new tables in the Noria
adapter.
This also adjust the table filter to be modified to add the new table
to the filter.
As a result, we know have one table filter object globally that is
passed as a reference to the per database adapter.
As of now, only one thread will be modifying/reading the table filter,
so we don't need to worry about synchronization.

Closes: #REA-4780

Release-Note-Core: Added support for snapshotting new tables without
  requiring a restart of the adapter. The new command accepts a comma
  list of tables:
  `ALTER READYSET ADD TABLES <db>.<table_name1>, <db>.<table_name2>`.

Change-Id: Ibfcaa1ac15d88133dddc70cafebd977d2d56b330
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8336
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>