adapter: UnsupportedSetMode for SET prepared statements
`UnsupportedSetMode` was implemented in the adapter for adhoc queries,
but never for prepared statements. As prepared statements have their
functionality spread across both `prepare` and `execute` in the
`adapter::Backend`, some complexity is necessarily introduced to
specially handle `SET` commands.

Additional care, around the close/`remove_statement()`, is
needed in the case of "no upstream connections".

Fixes: REA-4826
Fixes: #1381

Change-Id: I7ab35c9c0f1a98f42fe5cf7fab1ba008448c4c1c