adapter: Merge --allow-unsupported-set
`--allow-unsupported-set` was added in Ie7bfc9cd, and it allowed
readyset to naively ignore any `SET` commands it could not process.

About a year later, in I0005339c, we introduced the
`--unsupported-set-mode` flag and enum. One of the modes available is
`Allow`, that mapped directly onto the existing `--allow-unsupported-set`.

Confusingly, we did not remove the older flag, and we did not allow
setting `--unsupported-set-mode allow`.

Let's reduce our cli flag burden and just merge the two behaviors,
into the way it looks like it was intended to be.

Fixes: REA-4673
Fixes: #1354
Release-Note-Core: Remove hidden `--allow-unsupported-set` flag,
  replaced by `--unsupported-set-mode allow`.
Change-Id: I368ca7ee583bb64d3e44e22b81576c79551d62ff
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8063
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>