adapter: Default schema search path
When using readyset with no upstream database,
`--no-upsttream-connections`, we still need a correct schema search
path defined for postgres. We can't get the actual schema search path
as we can't connect to the upstream (at least not from the adapter),
thus this CL introduces a best-effort default for the schema search
path, based on the default of PG (`SHOW schema_path`). In practice
this will likely be more than sufficient as few operators change the
default search path.

Fixes: REA-4806
Fixes: #1377
Release-Note-Core: Set the schema search path to a standard default
  when using `--no-upstream-connections`.
Change-Id: I1464d51c09728055e10697bf7c8dc471ea272e9d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8065
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>