readyset-psql: expand extra_float_digits acceptable values
The postgres session parameter `extra_float_digits` "adjusts the
number of digits used for textual output of floating-point
values". [0]

The range of accetpable values is -15..3. As of postgres 12, values
greater than zero all output the "shortest-precise" format [1], so we
should accept all values greater than zero to be aliases of each
other. We previously only accepted the value of `1`, which was the
prior pg default.

This issue came up as the postgres driver, pgjdbc, always sends the command
`SET extra_float_digit = 3` on every new connection.

[0]
https://www.postgresql.org/docs/current/runtime-config-client.html
[1]
https://www.postgresql.org/docs/current/datatype-numeric.html

Fixes: REA-4824
Fixes: #1379
Release-Note-Core: Accept more values for the postgres session
  varaible `extra_float_digits`.
Change-Id: I160429cc01aa309ac3f7914c46aa1cbbd956083b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8066
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>