reader-map: Switch back to SmallVec on (write after) read
The optimization that used a B-tree to eliminate potential O(n^2)
startup time caused a performance regression when reading due to
repeated conversion from a B-tree back to a SmallVec, which was
unnecessary before the optimization.  That code is in the read path and
so must be extremely fast.

We add a third state to represent when we've started up and started
receiving reads.  In that state, we use a SmallVec, since we shouldn't
be getting too many writes.  The B-tree is now only used temporarily
when receiving initial writes.  The three states are:

(1) No reads, few writes
(2) No reads, many writes
(3) Reads received

Because we only have an immutable reference on reads, we can't fully
switch states at that time.  Instead, we use some interior mutability to
hold the result of the conversion, and we complete the state change on
the next write, when we have a mutable reference.  We switch to the new
third state on read regardless of whether we have gone through the
B-tree state.

Fixes: REA-4809
Change-Id: I61041b24bbebbab2c0450b6365b6de4d76f02d83
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8067
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>