Delete #[allow(clippy::indexing_slicing)]
This is a non-default lint that we explicitly enabled (at least in a
couple of places; it's not clear that some of these weren't totally
non-functional).  The lint worked so well for us that we had to disable
it with clippy turds 152 times.

If it's not working for us 152 times, maybe we shouldn't have turned
it on in the first place.

Change-Id: Ibd3eb5bfb0bd60acec34fab1b5addef8e8b227be
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8090
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>