readyset-dataflow: Rename publish functions to publish
At three levels of the stack, we variously called the operation
publish, refresh, and swap.  The bottom level's term is the best,
so that's what we're switching to.

("Publish" describes the effect.  "Refresh" connotes more significant
computation of data that aren't there and so is less correct.  "Swap"
exposes implementation details and doesn't say why you'd do it.)

Change-Id: Ia2fa3aabf3ab12a53b9ba2c83156e915d30fac84
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8098
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>