readyset-dataflow: Factor RequestPartialReplay packet into own struct
Using an enum with struct variants for network packet types is unwieldy
when you need to refer to a specific variant, which has no type name.
Thus, you can only pass around the outer type, and the inner type can
only be accessed by typing in a bunch of match boilerplate bloat.

Change-Id: I886638847db611367412b5fd9365f0b4095c3afb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8175
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>