readyset-dataflow: Factor Domain::upquery
We optimized the single-shard case (which is the only case for now) by
making the function about 50% longer to skip the use of a HashMap.  But
the HashMap's keys were small contiguous integers, and there is a
different data structure that is both very fast and ideally suited to
storing values keyed by small contiguous integers:  a vector.

Should the multi-shard code become live again, we've also reduced the
transient memory usage, to n from m * n plus HashMap overhead.

Change-Id: I561b3dcead62a53cfc19bff472f08ee4f054c300
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8216
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>