replicators: Skip replication events earlier
MySQL:
This patch introduces a new ReplicationAction called Empty. This is
used to skip replication events for tables that are not replicated.
This is done as soon as we identify the event to avoid unnecessary
processing of the event, like converting from the binlog to noria,
pass it back to the noria adapter to then be discarded.

Postgres:
We also skip the replication event as soon as we read
it from the WAL. Truncate is handled separately as it has a vector of
relation_ids.

Release-Note-Core: Improved the performance of replication by
  discarding unnecessary events earlier in the process.

Change-Id: Ieab0dd21526cf2545aee6e1294f85abfd15dc4f7