replicators: Skip replication events earlier
This patch introduces a new ReplicationAction called Empty. This is
used to skip replication events for tables that are not replicated.
This is done as soon as we identify the event to avoid unnecessary
processing of the event, like converting from the binlog to noria,
pass it back to the noria adapter to then be discarded.

Change-Id: Ieab0dd21526cf2545aee6e1294f85abfd15dc4f7