readyset-mysql: last-insert-id should default to 0
When an insert occurs on a mysql table with an auto-increment column
on the primary key, mysql's OK packet will return the id of the most
recent insert [0]. If not an insert or if the table did not have an
auto-increment column as the primary key, the value should be zero.

Readyset currently defaults to returning `1` as the value, which is
incorrect.

[0] https://dev.mysql.com/doc/dev/mysql-server/latest/
  page_protocol_basic_ok_packet.html

Change-Id: I874a840a2b100a3ac3643b0eb2651387fa0e5489
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8125
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>