forward COM_PING to backend connection
When the connection is idle in some proxies/connection pool's, it's
common to refresh the connection by sending a COM_PING(MySQL) or
select 1 query(Postgres). Currently we are not forwarding the COM_PING
to the backend connection, which can cause the connection to be closed
by the backend server.

Fixes: REA-4887
Fixes: #1387

Release-Note-Core: Fixes an issue that we misse to forward COM_PING to
  backend connection to keep the connection alive.
Change-Id: I92178a37d58f8336a669c2f259ca9a478b856973
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8112
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>