dataflow-state: Require lock to save meta in persistent state
While refactoring, I observed the locking here is kind of a mess.
We get a lock, modify something, drop it, then reacquire it shared
to read what we just did.

I have no idea whether the old code was correct or not, but I know
whether the new code is.  (It's better but not perfect.  I shall
continue to improve it presently.)

Change-Id: Iddbcfc66a81c84ae71c30f16d2165d6614224140