readyset-server: Remove table aliases on rewrite
When creating a cache, we perform a series of rewrites on the incoming
`SelectStatement`, almost of all of which are handled by
`SqlIncorporator::rewrite()`. We then perform an additional rewrite by
calling `SelectStatement::rewrite_table_aliases()`, which is a trait
defined in `readyset-sql-passes`, just like all the other
passes.

Unforuntely, no other code paths that call
`SqlIncorporator::rewrite()` also make the call to
`rewrite_table_aliases()`. The biggest problem is the Leader's
`view_names` endpoint, that the adapter calls to discover if a given
query is cached or not. It performs the standard rewrite passes, but
does not perform `rewrite_table_aliases()`, and hence misses when
trying to lookup into the `ExprRegistry`. The downside being that a
query that is actually cached is not thought to be cached, and thus
the adapter proxies to the upstream.

This CL adds the `rewrite_table_aliases()` to the standard set of
rewrites for `SelectStatement`. Also includes a clustertest to confirm
the functionality.

Fixes: REA-4897
Fixes: #1390
Release-Note-Core: Fix bug where cached queries that use table aliases
  are not found after restart.
Change-Id: I84d421e76c707590dd47e5943ad06b115cedf5e3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8116
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>