adapter: Lower log level for unsupported prepares
If a client application sends `SHOW` or other similar commands over
the extended query protocol (which we treat as "unnamed prepared
statements"), and readyset does not support that command, we'll just
proxy it to the upstream (if there is one). We used to log a message
at `warn`, without any relevant data about the query itself. If the
application repeatedly sends that command (for whatever legitimate
reason), we'll end up spamming the log file with no useful
information. This CL lowers the log level to `debug` as it's
ultimately not useful information when deployed in prod, and now also
prints the statement that was issued.

Change-Id: I3738f50c8d041c6103f56e6b4309be035144ac89