Add support for MySQL 5.7
This commits fixes some issues that prevent us from connecting and
replicating from a 5.7 instance. The main issue with 5.7 is the lack
of support for the `LOCK INSTACE FOR BACKUP` which allows tables to be
created during the backup.
During snapshot we do the following:
- get the list of tables
- for each table run a select 1 from table_name to add a metadata lock
  and prevent the table from being altered or dropped.
- get the current binlog position

If a new table happens to be created between the time we get the list
and before we get the current binlog position, we will not snapshot
the table. This is a rare case and the worst that can happen is the
table be missing from readyset in the first snapshot. If the user
restart readyset it will snapshot the table.

Fixes: REA-4910
Fixes: #1393

Release-Note-Core: Fixes issues that prevent us from connecting and
  replicating from a MySQL 5.7 instance.
Change-Id: I89b74edb9dd9a4ccfe00d55d0e32cd2efc491c54