dataflow-state: Metric for rocksdb block index and filters size
Output a new metric to capture, as a gauge, the current memory use by
rocksdb block index and filters. These are loaded from each sstable in
the rocksdb instance, and are held in memory for the lifetime of the
sstable.

The correct way to get a reasonable value is to probe each active
(loaded) column family in the rocksdb instance for its size, sum up
all values, and emit to the metric.

Addresses: REA-4895
Change-Id: I917d8810a365da73d50d35eb018b79acf8c4bcde
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8142
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>