dataflow-state: Reduce sstable size
A long time ago, when rocksdb was first added to Noria, the
`target_file_size_base` property of rocksdb was set to
256 MiB (35259d7734dfd60df44996450f04584f3f395d12). This value is
significantly higher than the default, 64 MiB, and no reasoning is
provided for the choice.

When files are larger, they may be pulled into more compaction
activity as they will span an incrementally larger portion of the key
range (at a given level).

This CL removes the custom value an lieu of the default of 64 MiB.

Addresses: REA-4895
Change-Id: I758ae90b8100f142f8dabb779875716280b5840a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8144
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>