dataflow-state: Raise sstable index block size
The default block size for sstables is 4 KiB. With a smaller
block size, that means there will be more corresponding entries in the
index for those blocks. When there are many sstables in a given
rocksdb instance, the loaded index blocks can take a lot of
memory space. Thus, by raising the size of the blocks, there will be fewer
blocks and thus less memory consumed. In the rocksdb docs [0], it is
stated:

"By default our block size is 4KB, although we usually run with
16-32KB in production."

The trade off, of course, is that on the read path, once we find the
correct block to read from, we'll need to do a bit more scanning to
find the key (or not, if we've hit a false positive). However, the
trade off is reasonable since we can free many GiBs of memory on a
large rocksdb database, and use that memory for more data in the
readyset caches.

This CL takes the linked advice and raises the block size to 32 KiB.

[0] https://github.com/facebook/rocksdb/wiki/ \
  Memory-usage-in-RocksDB#indexes-and-filter-blocks

Addresses: REA-4895
Change-Id: Ie7ab4bbd1fa1f3cb28782608705564ccd1d4a64a