Replace serde_yaml with serde_yaml_ng
We were on a rather old version of more than two years ago.  The latest
version is deprecated, and it has a dependency that is unmaintained.

The replacement is semi-compatible.  The API requires a small change
that was probably required by a new version of the old crate that we
weren't using.  The YAML accepted appears to have changed as well with
respect to enum handling:  the YAML now needs to have "!variant" instead
of "variant:".  I attempted to fix this up, but many of the YAML files
didn't work before the change.  I have gotten build tests to pass, but
some fallout in the form of nightly failure will not be wholly
unexpected.

Change-Id: I7de0c6ef72ff16d84d9a0b61fc8663f3615093c1