adapter: Change query metric's `database_type` tag
In the `QueryLogger`, when we proxy a query to the upstream database,
we set the metric's `database_type` tag to `mysql` even if the the
upstream database is something else (that is, postgres). This CL
changes that value to a more generic `upstream` value (to
differentiate it from `readyset`).

Fixes: REA-4928
Release-Note-Core: Change query metric's `database_type` tag value to
  `upstream` when serving from an upstream database.
Change-Id: I5544fd5a31fd8fa0f9f122511779f68c9b7e737c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8210
Tested-by: Buildkite CI
Reviewed-by: Sandaruwan Gunathilake <sandaruwan@readyset.io>