Remove default timeout for query migration
We used to fail by default cache creation that took longer than 30
minutes.  This duration could occur if the cache creation invoked the
creation of 1 or more large secondary indices.

Previously, we'd still finish the index creation in the background, but
the cache would not actually be created.  A second invocation of the
cache creation could then complete almost instantly within the typically
generous 30 minutes, yielding the desired cache.

Now, we remove the default timeout.  Cache creation that takes longer
than 30 minutes should succeed with a single invocation.

The migration request timeout (--migration-request-timeout-ms), and a
related controller request timeout (--controller-request-timeout-ms),
can now be set to 0 to specify an unlimited timeout.

Fixes: REA-4894
Change-Id: Ie50610fdc31b761de44cd178c0e2adf488056438
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8230
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI