data-generator: Use actual PRNG when verifying random_value_of_type
Generating MySQL MEDIUMINT values (effectively 24 bits) can take
unreasonably long in certain circumstances, e.g. if `initial` is greater
than 2**24 and `increment` is small, as this will cause
`StepRng::next_u32` to continuously generate values outside the 24 bit
range, causing `Rng::gen_range` to keep sampling trying to find one in
its range. Using `SmallRng` is not noticeably slower for the 1024
proptest cases we'll do in each of these runs and will actually generate
pseudorandom numbers and not effectively hang like this.

Note that although the doc comments warn not to rely on `SmallRng` for
reproducibility due to platform differences and possible library
changes, all 64 bit platforms currently use `Xoshiro256PlusPlus` and we
can just switch to use that crate directly (`rand_xoshiro`) if it
changes and we want to debug a failure that happened on a previous
version of the `rand` crate but we want to do so while using a new
version of the `rand` crate.

Change-Id: If8c095d5da0f248a63adea0b7631eb2a25e6abf3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8233
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>