adapter: Show upstream execution counts in SHOW PROXIED QUERIES
32239b89 replaced "mysql" with "upstream" in the metrics, causing
executions to not be snapshotted and reported. We don't care about the
distinction between MySQL and Postgres here, even though the metric is
called "database_type"; just whether it's upstream or not. So arguably
this label should be renamed away from "database_type", but I'm not sure
where else we might be depending on that label name. Elsewhere we have
an "upstream_db_type" label where we might want to reinstate an enum for
the distinction between MySQL and Postgres, but I don't think it is that
important.

Change-Id: I24a11acdec8d6a4a8fa141486db44eaf9ade3e1c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8245
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>