readyset-server: Fix erroneously exited span in domain loop
Apparently several people didn't know that `let _ = ...` is a bit of a
land mine in that the value is dropped immediately, unlike other
bindings with a leading underscore.  In the case of a guard object, that
is undesired behavior.

Additionally, Span::enter is incorrect in async code that contains an
await point.  (At least the documentation alerts you to that.)

Change-Id: Ia4c10f4ee66c88fba590ed74acc96ef0192f61a2