grafana: Upgrade grafana tests
This commit upgrades grafana tests to test two specific queries.
One cached and one proxied. The screenshots are uploaded as
artifacts to buildkite. The HTML output is removed from the
debug log to reduce the noise and added as artifacts.

This also fixes the broken query specific dashboard for proxied
queries.

Change-Id: I6e58230b06df46a177a1657762c88f50bce771f3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8283
Reviewed-by: Ron Hough <ron@readyset.io>
Tested-by: Buildkite CI