Display Domain id in EXPLAIN MATERIALIZATIONS
Often we need to track down which domain is assigned to a specific
materialization (base table, full mat, partial mat node). At the
moment, we need to generate a full graphviz representation of the
dataflow graph and scan it manually to find the domain id. This is
cumbersome and in cases where we have multiple caches this can be
confusing.
This commit adds the domain_index in the EXPLAIN MATERIALIZATIONS
output so that we can easily track down the domain id of a specific
materialization.

Closes: #REA-5003

Change-Id: I5e687f1cb6a2aa1ed21a78e528419020272b9862