dataflow: Allow domain packet failpoint to fail a specified domain
This is an abuse of the failpoint "return" API. It is intended to be
used to force the enclosing function to return a specified value.
Instead, we use it to determine whether this failpoint should trigger
based on interpreting the provided "return value" as a domain ID.

For example, to make domain 4 return an error:

```sh
cargo run --bin failpoint -- --controller-address \
    http://localhost:6033 handle-packet "return(4)"
```

Currently, it causes the domain to return a `ReadySetError::Internal`
error. Future work could re-interpret the "return value" using the same
logic as the failpoint crate (e.g. `return(4,panic)` to panic domain 4).
And we could package this up as a reusable macro which inlines the
"returning" inner function. But if we find we need this functionality in
that many more places, we should consider modifying or replacing the
failpoint crate instead.

Change-Id: I1e2c79c334c8f5ccebddfbc4fb52dca9dae6752a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8358
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>