Remove the need for exact row count
This commit removes the requirement for MySQL snapshot for requiring
the exact number of rows it will snapshot in advance. We use the
number of rows to show snapshot progress and also to know for sure
that we have snapshoted the required number of rows. Querying the
table to get the number of rows in MySQL requires a full index scan.
This can be time consuming for large tables.
We don't necessarlly need to know the exact number of tables. In a
server with up to date table statistics the estimation of table rows
will be pretty much accurate.

Fixes: #REA-4886
Closes: #1386

Release-Note-Core: Improve MySQL snapshot performance by removing the
  need for exact row count for snapshot.
Change-Id: Idf887a0b8abb8e20113b3585af49c75d3da8b05f