Suppress warning for MySQL rotate event.
When we receive a MySQL ROTATE_EVENT we send a
ReplicationAction::LogPosition to update the offset of all tables to
the new position as it's guaranteed that we don't have any missing
event in the previous binlog for any table.
The problem is that our driver intercepts the first event in the
binlog as a ROTATE_EVENT and we end up sending a new rotate event
with the same position and log a warning in the Noria Adapter as we
should always process an event with a higher position than the
previous one.

This commits adjust how we handle the ROTATE_EVENT in the MySQL
replication to avoid sending a new ReplicationAction::LogPosition
when we receive a ROTATE_EVENT with the same binlog as the previous
one and we are processing the position 0 of binlog.

Change-Id: I4b1b2c5bb89f65c54548145140efb67573de8050
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/8378
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>
Reviewed-by: Jason Brown <jason.b@readyset.io>