non-sql: propagate default collation
In preparation to handle case-insensitive text in MySQL, we need to
propagate the default collation to the column definition.
MySQL collation works on a per-column basis, if you omit the collation
from the column definition, the default collation is used.

Depending where we are in the code, like when we are converting from
SqlType to DfType, we don't have access to the table definition, where
we would extract the default collation from.

Change-Id: I8b64075c2a7059c10afda87e41cf4db6f6b25433